+++ test/unit/compiler.spec.js
@@ -64,17 +64,17 @@ describe("Peggy compiler", () => {
     expect(ast).to.be.an("object");
     expect(() => compiler.compile(ast, compiler.passes, {
       output: "source-and-map",
-    })).to.throw("Must provide grammarSource (as a string) in order to generate source maps");
+    })).to.throw("Must provide grammarSource (as a string or GrammarLocation) in order to generate source maps");
     expect(() => compiler.compile(ast, compiler.passes, {
       output: "source-and-map",
       grammarSource: "",
-    })).to.throw("Must provide grammarSource (as a string) in order to generate source maps");
+    })).to.throw("Must provide grammarSource (as a string or GrammarLocation) in order to generate source maps");
     // Don't run on old IE
     if (typeof TextEncoder === "function") {
       expect(() => compiler.compile(ast, compiler.passes, {
         output: "source-with-inline-map",
         grammarSource: { toString() { return ""; } },
-      })).to.throw("Must provide grammarSource (as a string) in order to generate source maps");
+      })).to.throw("Must provide grammarSource (as a string or GrammarLocation) in order to generate source maps");
     }
   });
 });